Move `OtherSession` test helper to rustls-test
