+++ rustls/tests/api.rs
@@ -4,7 +4,6 @@
 
 use std::fmt::Debug;
 use std::io::{self, BufRead, IoSlice, Read, Write};
-use std::ops::{Deref, DerefMut};
 use std::sync::Mutex;
 use std::sync::atomic::{AtomicUsize, Ordering};
 use std::{fmt, mem};
@@ -21,9 +20,9 @@ use rustls::internal::msgs::message::{Message, MessagePayload, PlainMessage};
 use rustls::server::{ClientHello, ParsedCertificate, ResolvesServerCert};
 use rustls::{
     AlertDescription, ApiMisuse, CertificateError, CertificateIdentity, CipherSuite, ClientConfig,
-    ClientConnection, ConnectionCommon, ConnectionTrafficSecrets, ContentType, DistinguishedName,
-    Error, ExtendedKeyPurpose, HandshakeKind, HandshakeType, InconsistentKeys, InvalidMessage,
-    KeyLog, KeyingMaterialExporter, NamedGroup, PeerIdentity, PeerIncompatible, PeerMisbehaved,
+    ClientConnection, ConnectionTrafficSecrets, ContentType, DistinguishedName, Error,
+    ExtendedKeyPurpose, HandshakeKind, HandshakeType, InconsistentKeys, InvalidMessage, KeyLog,
+    KeyingMaterialExporter, NamedGroup, PeerIdentity, PeerIncompatible, PeerMisbehaved,
     ProtocolVersion, RootCertStore, ServerConfig, ServerConnection, SideData, SignatureScheme,
     Stream, StreamOwned, SupportedCipherSuite, sign,
 };
@@ -2145,129 +2144,6 @@ fn buf_read() {
     assert_eq!(reader.fill_buf().unwrap(), b"world");
     reader.consume(5);
     check_fill_buf_err(&mut reader, io::ErrorKind::WouldBlock);
-}
-
-struct OtherSession<'a, C, S>
-where
-    C: DerefMut + Deref<Target = ConnectionCommon<S>>,
-    S: SideData,
-{
-    sess: &'a mut C,
-    pub reads: usize,
-    pub writevs: Vec<Vec<usize>>,
-    fail_ok: bool,
-    pub short_writes: bool,
-    pub last_error: Option<rustls::Error>,
-    pub buffered: bool,
-    buffer: Vec<Vec<u8>>,
-}
-
-impl<'a, C, S> OtherSession<'a, C, S>
-where
-    C: DerefMut + Deref<Target = ConnectionCommon<S>>,
-    S: SideData,
-{
-    fn new(sess: &'a mut C) -> OtherSession<'a, C, S> {
-        OtherSession {
-            sess,
-            reads: 0,
-            writevs: vec![],
-            fail_ok: false,
-            short_writes: false,
-            last_error: None,
-            buffered: false,
-            buffer: vec![],
-        }
-    }
-
-    fn new_buffered(sess: &'a mut C) -> OtherSession<'a, C, S> {
-        let mut os = OtherSession::new(sess);
-        os.buffered = true;
-        os
-    }
-
-    fn new_fails(sess: &'a mut C) -> OtherSession<'a, C, S> {
-        let mut os = OtherSession::new(sess);
-        os.fail_ok = true;
-        os
-    }
-
-    fn flush_vectored(&mut self, b: &[io::IoSlice<'_>]) -> io::Result<usize> {
-        let mut total = 0;
-        let mut lengths = vec![];
-        for bytes in b {
-            let write_len = if self.short_writes {
-                if bytes.len() > 5 {
-                    bytes.len() / 2
-                } else {
-                    bytes.len()
-                }
-            } else {
-                bytes.len()
-            };
-
-            let l = self
-                .sess
-                .read_tls(&mut io::Cursor::new(&bytes[..write_len]))?;
-            lengths.push(l);
-            total += l;
-            if bytes.len() != l {
-                break;
-            }
-        }
-
-        let rc = self.sess.process_new_packets();
-        if !self.fail_ok {
-            rc.unwrap();
-        } else if rc.is_err() {
-            self.last_error = rc.err();
-        }
-
-        self.writevs.push(lengths);
-        Ok(total)
-    }
-}
-
-impl<C, S> io::Read for OtherSession<'_, C, S>
-where
-    C: DerefMut + Deref<Target = ConnectionCommon<S>>,
-    S: SideData,
-{
-    fn read(&mut self, mut b: &mut [u8]) -> io::Result<usize> {
-        self.reads += 1;
-        self.sess.write_tls(b.by_ref())
-    }
-}
-
-impl<C, S> io::Write for OtherSession<'_, C, S>
-where
-    C: DerefMut + Deref<Target = ConnectionCommon<S>>,
-    S: SideData,
-{
-    fn write(&mut self, _: &[u8]) -> io::Result<usize> {
-        unreachable!()
-    }
-
-    fn flush(&mut self) -> io::Result<()> {
-        if !self.buffer.is_empty() {
-            let buffer = mem::take(&mut self.buffer);
-            let slices = buffer
-                .iter()
-                .map(|b| io::IoSlice::new(b))
-                .collect::<Vec<_>>();
-            self.flush_vectored(&slices)?;
-        }
-        Ok(())
-    }
-
-    fn write_vectored(&mut self, b: &[io::IoSlice<'_>]) -> io::Result<usize> {
-        if self.buffered {
-            self.buffer
-                .extend(b.iter().map(|s| s.to_vec()));
-            return Ok(b.iter().map(|s| s.len()).sum());
-        }
-        self.flush_vectored(b)
-    }
 }
 
 #[test]
